break down class
